requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
emoji==0.5.0
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
nltk==3.3
numpy==1.15.0
oauthlib==2.1.0
pandas==0.23.4
Pillow==5.2.0
pprint==0.1
pyclustering==0.8.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-twitter==3.4.2
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
tqdm==4.24.0
urllib3==1.23
virtualenv==16.0.0
Werkzeug==0.14.1
